Type changes
The type of variables and result is changed. Provides the division with a fractional part in result.

+++ calculator/calcmodel.cpp
@@ -114,7 +114,7 @@ QString CalcModel::getCurrentValue()
     return QString::number(_currentValue);
 }
 
-int CalcModel::calculator(int x, int y, QString inputOp)
+float CalcModel::calculator(float x, float y, QString inputOp)
 {
     switch(inputOp.toStdString()[0])
         {
